requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
faiss-cpu==1.13.2
fastapi==0.129.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
ollama==0.6.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pypdf==6.7.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
